requirements.txt
pysoem==1.1.9
pytango==9.3.6